Make TunnelEndPointInfo work on Uint64
Rather than forcing Uint64->String conversion and subsequent
String->Uint64 conversions, make TunnelEndPointInfo store Uint64s
and lazily derive (and cache) String representation.

+++ itm/itm-impl/src/main/java/org/opendaylight/genius/itm/rpc/ItmManagerRpcService.java
@@ -1172,9 +1172,8 @@ public class ItmManagerRpcService implements ItmRpcService {
 
         Optional<DPNTEPsInfo> dpntePsInfoOptional = dpnTEPsInfoCache.get(InstanceIdentifier.builder(DpnEndpoints.class)
                 .child(DPNTEPsInfo.class, new DPNTEPsInfoKey(
-                    // FIXME: the cache should be caching this value, not just as a String
-                    Uint64.valueOf(dpnTepStateCache.getTunnelEndPointInfoFromCache(interfaceInfo.getTunnelName())
-                        .getDstEndPointInfo())))
+                    dpnTepStateCache.getTunnelEndPointInfoFromCache(interfaceInfo.getTunnelName())
+                        .getDstEndPointInfo()))
                 .build());
         Integer dstId;
         if (dpntePsInfoOptional.isPresent()) {
